+++ buession-geoip/src/main/java/com/buession/geoip/AbstractResolver.java
@@ -21,7 +21,7 @@
  * +------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										|
  * | Author: Yong.Teng <webmaster@buession.com> 													|
- * | Copyright @ 2013-2022 Buession.com Inc.														|
+ * | Copyright @ 2013-2023 Buession.com Inc.														|
  * +------------------------------------------------------------------------------------------------+
  */
 package com.buession.geoip;
@@ -44,6 +44,10 @@ import java.util.Locale;
  * @author Yong.Teng
  */
 public abstract class AbstractResolver implements Resolver {
+
+	public final static String DEFAULT_CITY_DB = "/maxmind/City.mmdb";
+
+	public final static String DEFAULT_ASN_DB = "/maxmind/ASN.mmdb";
 
 	private final static Logger logger = LoggerFactory.getLogger(AbstractResolver.class);