+++ buession-geoip/src/main/java/com/buession/geoip/spring/GeoIPResolverFactory.java
@@ -19,17 +19,18 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 											   |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.geoip.spring;
 
 import com.buession.core.utils.Assert;
+import com.buession.geoip.DatabaseResolver;
 import com.buession.geoip.Resolver;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Path;
@@ -40,8 +41,6 @@ import java.nio.file.Path;
  * @author Yong.Teng
  */
 public class GeoIPResolverFactory {
-
-	protected final static String DEFAULT_CITY_DB = "maxmind/City.mmdb";
 
 	/**
 	 * IP 库文件路径
@@ -54,8 +53,23 @@ public class GeoIPResolverFactory {
 	private InputStream stream;
 
 	/**
+	 * ASN 库文件路径
+	 *
+	 * @since 2.2.0
+	 */
+	private File asnDbPath;
+
+	/**
+	 * ASN 库文件流
+	 *
+	 * @since 2.2.0
+	 */
+	private InputStream asnStream;
+
+	/**
 	 * IP 库加载模式
 	 */
+	@Deprecated
 	private LoadMode loadMode = LoadMode.STREAM;
 
 	/**
@@ -67,11 +81,13 @@ public class GeoIPResolverFactory {
 	 * 返回 IP 库文件路径
 	 *
 	 * @return IP 库文件路径
+	 *
+	 * @throws IOException
+	 * 		数据库文件不存在时
 	 */
 	public File getDbPath() throws IOException{
 		if(dbPath == null){
-			ClassPathResource resource = new ClassPathResource(DEFAULT_CITY_DB);
-			dbPath = resource.getFile();
+			dbPath = new File(DatabaseResolver.class.getResource(DatabaseResolver.DEFAULT_CITY_DB).getFile());
 		}
 
 		return dbPath;
@@ -84,13 +100,11 @@ public class GeoIPResolverFactory {
 	 * 		IP 库文件 {@link Resource}
 	 *
 	 * @throws IOException
-	 * 		IO 错误
+	 * 		数据库文件不存在时
 	 */
 	public void setDbPath(Resource dbPath) throws IOException{
 		Assert.isNull(dbPath, "Ip database path cloud not be null.");
-
 		this.dbPath = dbPath.getFile();
-		this.loadMode = LoadMode.FILE;
 	}
 
 	/**
@@ -100,13 +114,11 @@ public class GeoIPResolverFactory {
 	 * 		IP 库文件路径
 	 *
 	 * @throws IOException
-	 * 		IO 错误
+	 * 		数据库文件不存在时
 	 */
 	public void setDbPath(File dbPath) throws IOException{
 		Assert.isNull(dbPath, "Ip database path cloud not be null.");
-
 		this.dbPath = dbPath;
-		this.loadMode = LoadMode.FILE;
 	}
 
 	/**
@@ -116,13 +128,11 @@ public class GeoIPResolverFactory {
 	 * 		IP 库文件路径
 	 *
 	 * @throws IOException
-	 * 		IO 错误
+	 * 		数据库文件不存在时
 	 */
 	public void setDbPath(Path dbPath) throws IOException{
 		Assert.isNull(dbPath, "Ip database path cloud not be null.");
-
 		this.dbPath = dbPath.toFile();
-		this.loadMode = LoadMode.FILE;
 	}
 
 	/**
@@ -132,7 +142,7 @@ public class GeoIPResolverFactory {
 	 * 		IP 库文件路径
 	 *
 	 * @throws IOException
-	 * 		IO 错误
+	 * 		数据库文件不存在时
 	 */
 	public void setDbPath(String dbPath) throws IOException{
 		Assert.isBlank(dbPath, "Ip database path cloud not be null or empty.");
@@ -140,14 +150,93 @@ public class GeoIPResolverFactory {
 	}
 
 	/**
+	 * 返回 ASN 库文件路径
+	 *
+	 * @return ASN 库文件路径
+	 *
+	 * @throws IOException
+	 * 		ASN 库文件不存在时
+	 * @since 2.2.0
+	 */
+	public File getAsnDbPath() throws IOException{
+		if(asnDbPath == null){
+			asnDbPath = new File(DatabaseResolver.class.getResource(DatabaseResolver.DEFAULT_ASN_DB).getFile());
+		}
+
+		return asnDbPath;
+	}
+
+	/**
+	 * 设置 ASN 库文件路径
+	 *
+	 * @param asnDbPath
+	 * 		ASN 库文件 {@link Resource}
+	 *
+	 * @throws IOException
+	 * 		ASN 库文件不存在时
+	 * @since 2.2.0
+	 */
+	public void setAsnDbPath(Resource asnDbPath) throws IOException{
+		Assert.isNull(asnDbPath, "Ip asn database path cloud not be null.");
+		this.asnDbPath = asnDbPath.getFile();
+	}
+
+	/**
+	 * 设置 ASN 库文件路径
+	 *
+	 * @param asnDbPath
+	 * 		IP 库文件路径
+	 *
+	 * @throws IOException
+	 * 		ASN 库文件不存在时
+	 * @since 2.2.0
+	 */
+	public void setAsnDbPath(File asnDbPath) throws IOException{
+		Assert.isNull(asnDbPath, "Ip asn database path cloud not be null.");
+		this.asnDbPath = asnDbPath;
+	}
+
+	/**
+	 * 设置 ASN 库文件路径
+	 *
+	 * @param asnDbPath
+	 * 		ASN 库文件路径
+	 *
+	 * @throws IOException
+	 * 		数据库文件不存在时
+	 * @since 2.2.0
+	 */
+	public void setAsnDbPath(Path asnDbPath) throws IOException{
+		Assert.isNull(asnDbPath, "Ip asn database path cloud not be null.");
+		this.asnDbPath = asnDbPath.toFile();
+	}
+
+	/**
+	 * 设置 ASN 库文件路径
+	 *
+	 * @param asnDbPath
+	 * 		ASN 库文件路径
+	 *
+	 * @throws IOException
+	 * 		数据库文件不存在时
+	 * @since 2.2.0
+	 */
+	public void setAsnDbPath(String asnDbPath) throws IOException{
+		Assert.isBlank(asnDbPath, "Ip asn database path cloud not be null or empty.");
+		setAsnDbPath(new File(asnDbPath));
+	}
+
+	/**
 	 * 返回 IP 库文件流
 	 *
 	 * @return IP 库文件流
+	 *
+	 * @throws IOException
+	 * 		数据库文件不存在时
 	 */
 	public InputStream getStream() throws IOException{
 		if(stream == null){
-			ClassPathResource resource = new ClassPathResource(DEFAULT_CITY_DB);
-			stream = resource.getInputStream();
+			stream = DatabaseResolver.class.getResourceAsStream(DatabaseResolver.DEFAULT_CITY_DB);
 		}
 
 		return stream;
@@ -161,9 +250,37 @@ public class GeoIPResolverFactory {
 	 */
 	public void setStream(InputStream stream){
 		Assert.isNull(stream, "Ip database stream cloud not be null.");
-
 		this.stream = stream;
-		this.loadMode = LoadMode.STREAM;
+	}
+
+	/**
+	 * 返回 ASN 库文件流
+	 *
+	 * @return ASN 库文件流
+	 *
+	 * @throws IOException
+	 * 		ASN 库文件不存在时
+	 * @since 2.2.0
+	 */
+	public InputStream getAsnStream() throws IOException{
+		if(asnStream == null){
+			asnStream = DatabaseResolver.class.getResourceAsStream(DatabaseResolver.DEFAULT_ASN_DB);
+		}
+
+		return asnStream;
+	}
+
+	/**
+	 * 设置 asnStream 库文件流
+	 *
+	 * @param asnStream
+	 * 		asnStream 库文件流
+	 *
+	 * @since 2.2.0
+	 */
+	public void setAsnStream(InputStream asnStream){
+		Assert.isNull(asnStream, "Ip asn database stream cloud not be null.");
+		this.asnStream = asnStream;
 	}
 
 	/**
@@ -171,6 +288,7 @@ public class GeoIPResolverFactory {
 	 *
 	 * @return IP 库加载模式
 	 */
+	@Deprecated
 	public LoadMode getLoadMode(){
 		return loadMode;
 	}