+++ buession-geoip/src/test/java/com/buession/geoip/DatabaseResolverTest.java
@@ -27,6 +27,7 @@ package com.buession.geoip;
 import com.buession.geoip.model.Country;
 import com.buession.geoip.model.District;
 import com.buession.geoip.model.Location;
+import com.buession.geoip.spring.GeoIPResolverFactoryBean;
 import org.junit.Test;
 
 /**
@@ -46,6 +47,25 @@ public class DatabaseResolverTest {
 	@Test
 	public void country1() throws Exception{
 		DatabaseResolver resolver = new DatabaseResolver(
+				DatabaseResolver.class.getResourceAsStream("/maxmind/City.mmdb"));
+		Country country = resolver.country(3739974408L);
+		System.out.println(country);
+	}
+
+	@Test
+	public void country2() throws Exception{
+		GeoIPResolverFactoryBean factoryBean = new GeoIPResolverFactoryBean();
+
+		factoryBean.afterPropertiesSet();
+
+		DatabaseResolver resolver = factoryBean.getObject();
+		Country country = resolver.country(3739974408L);
+		System.out.println(country);
+	}
+
+	@Test
+	public void country3() throws Exception{
+		DatabaseResolver resolver = new CacheDatabaseResolver(
 				DatabaseResolver.class.getResourceAsStream("/maxmind/City.mmdb"));
 		Country country = resolver.country(3739974408L);
 		System.out.println(country);